Regular upgrade[3.2.0-SNAPSHOT]

+++ albedo-common/albedo-common-core/src/main/java/com/albedo/java/common/core/vo/ScheduleVo.java
@@ -37,28 +37,28 @@ public class ScheduleVo implements Serializable {
 
 	private ScheduleConstants.MessageType messageType;
 
-	private Integer jobId;
+	private Long jobId;
 
 	private String jobGroup;
 
 	private String data;
 
-	public ScheduleVo(ScheduleConstants.MessageType messageType, Integer jobId, String jobGroup) {
+	public ScheduleVo(ScheduleConstants.MessageType messageType, Long jobId, String jobGroup) {
 		this.messageType = messageType;
 		this.jobId = jobId;
 		this.jobGroup = jobGroup;
 	}
 
-	public ScheduleVo(ScheduleConstants.MessageType messageType, Integer jobId) {
+	public ScheduleVo(ScheduleConstants.MessageType messageType, Long jobId) {
 		this.messageType = messageType;
 		this.jobId = jobId;
 	}
 
-	public static ScheduleVo create(ScheduleConstants.MessageType messageType, Integer jobId, String jobGroup) {
+	public static ScheduleVo create(ScheduleConstants.MessageType messageType, Long jobId, String jobGroup) {
 		return new ScheduleVo(messageType, jobId, jobGroup);
 	}
 
-	public static ScheduleVo create(ScheduleConstants.MessageType messageType, Integer jobId) {
+	public static ScheduleVo create(ScheduleConstants.MessageType messageType, Long jobId) {
 		return new ScheduleVo(messageType, jobId);
 	}
 
@@ -66,19 +66,19 @@ public class ScheduleVo implements Serializable {
 		return new ScheduleVo(messageType, null, null, data);
 	}
 
-	public static ScheduleVo createPause(Integer jobId, String jobGroup) {
+	public static ScheduleVo createPause(Long jobId, String jobGroup) {
 		return create(ScheduleConstants.MessageType.PAUSE, jobId, jobGroup);
 	}
 
-	public static Object createResume(Integer jobId, String jobGroup) {
+	public static Object createResume(Long jobId, String jobGroup) {
 		return create(ScheduleConstants.MessageType.RESUME, jobId, jobGroup);
 	}
 
-	public static Object createDelete(Integer jobId, String jobGroup) {
+	public static Object createDelete(Long jobId, String jobGroup) {
 		return create(ScheduleConstants.MessageType.DELETE, jobId, jobGroup);
 	}
 
-	public static Object createRun(Integer jobId, String jobGroup) {
+	public static Object createRun(Long jobId, String jobGroup) {
 		return create(ScheduleConstants.MessageType.RUN, jobId, jobGroup);
 	}
